apps/plgrzr-trainer/requirements.txt

numpy>=1.21.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
python-doctr>=0.6.0
easyocr>=1.6.2
pix2tex>=0.4.0
Pillow>=9.0.0
tensorflow>=2.8.0
PyMuPDF>=1.19.0
albumentations>=1.1.0
shapely>=1.8.0
scipy>=1.7.0
matplotlib>=3.4.0
tqdm>=4.62.0
weasyprint>=54.0
unidecode>=1.3.0
transformers>=4.18.0
requests>=2.27.0
sentencepiece>=0.1.96
python-Levenshtein>=0.12.2
pandas>=1.4.0